\name{Textual Output for Various Functions}
\alias{chiSquareWords}
\alias{KruskalWallisTest2}
\alias{confintBinomialWords}
\alias{confintBinomial2}
\alias{confintContinuousWords}
\alias{confintContinuous2}
\alias{generalizedLinearModel2}
\alias{independentSamplesTTestWords}
\alias{independentSamplesTTest2}
\alias{kruskalWallisWords}
\alias{linearModel2}
\alias{logitWords}
\alias{multipleRegressionWords}
\alias{ologitWords}
\alias{oneWayAnova2}
\alias{ordinalRegressionModelOrdinal2}
\alias{pairedTTestWords}
\alias{pairedTTest2}
\alias{pairedWilcoxonTest2}
\alias{singleTTestWords}
\alias{singleProportionTestWords}
\alias{singleProportionTest2}
\alias{singleSampleTTest2}
\alias{twoSampleProportionsTestWords}
\alias{twoSampleProportionsTest2}
\alias{twoSampleWilcoxonTest2}
\alias{twoWayTable2}
\alias{wilcoxonMannWhitneyWords}
\alias{wilcoxonWords}
\alias{wordsAnova}
\alias{resetGLM}
\alias{resetLinearModel}

\title{Textual Interpretations: MPA Stats}
\description{
  Print textual interpretations of various tests in the RcmdrPlugin.MPAStats package. 
  These methods are not intended to be called directly by the user via the command line.
  These methods will be called via the R Commander console
  and will produce appropriate output and textual interpretations of selected analyses. 
}
\author{
  Richard Payne <\email{paynedrichard@gmail.com}>
}
