\name{RcmdrPlugin.BYUMPA-package}
\alias{RcmdrPlugin.BYUMPA-package}
\alias{RcmdrPlugin.BYUMPA}
\docType{package}
\title{ BYUMPA Plugin for Rcmdr }
\description{
This package is an R Commander plugin that accompanies and enhances 
BYUMPA, a separate library that contains a collection of functions 
made for public management and policy analysis statistics.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.BYUMPA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-19\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Andrew Heiss

Maintainer: Andrew Heiss <andrew@andrewheiss.com>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }

\examples{

}
