\name{RcmdrPlugin.MPAStats-internal}
\title{Internal RcmdrPlugin.MPAStats Objects}
\alias{checkStructure}
\alias{confintBinomial}
\alias{confintContinuous}
\alias{confintPoisson}
\alias{factorChange}
\alias{ordinalRegressionModel.ordinal}
\alias{resetclm}
\alias{sclmP}
\alias{summarizeDataframe}
\alias{DistWords}
\alias{DistWords2}
\alias{distdiagnostics}
\alias{distnormal}
\description{Internal RcmdrPlugin.MPAStats objects.}
\details{These are not to be called by the user.}
\keyword{internal}
