\name{RcmdrPlugin.MPAStats-package}
\alias{RcmdrPlugin.MPAStats-package}
\alias{RcmdrPlugin.MPAStats}
\docType{package}
\title{ R Commander Plug-in for MPA Statistics }
\description{
This package provides an R Commander plugin that offers a 
unified menu of new and pre-existing statistical functions 
related to public management and policy analysis statistics.
Functions and menus have been renamed according to the usage
in PMGT 630 in the Master of Public Administration program at
Brigham Young University.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.MPAStats\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2013-06-06\cr
License: \tab GPL (>2)\cr
}

The plug-in adds an additional top-level menu named \strong{MPA Statistics} to 
the \emph{R Commander} interface, with the following submenus: 
\strong{Descriptive statistics}, \strong{Single-sample statistics}, 
\strong{Paired value testing}, \strong{Bivariate analysis} (with additional submenus 
that divide statistical tests by the level of measurement of the dependent and 
independent variables), and \strong{Multivariate analysis}. Additionally, there is 
a function to calculate logistic regression factor change coefficients (e^b) at 
\strong{Models -> Factor change coefficients}.
}
\author{
Andrew Heiss, Richard Payne

Maintainer: Christa Schank <christa.schank@gmail.com>
}

\keyword{ package }
